Sea animal admin panel structure change
- Seperate data list and form to two pages
